convert web client to React + Redux
At the moment, this removes the ability to set queue order. Also, the
song input box does not disable while waiting, which could lead to
glitches in the queue list if multiple songs are submitted before they
have finished processing.

This updates discord.js to the master branch, which is currently the
upcoming v12.